requirements.txt
paho-mqtt
etcd3
dotenv-settings-handler
python-dotenv
